refactored some folders to use an index file; added new utility file
