+++ datasources/index.js
@@ -1,0 +1,7 @@
+const UserAPI = require('./UserAPI');
+const RulesAPI = require('./RulesAPI');
+
+module.exports = {
+  UserAPI,
+  RulesAPI,
+};
