+++ models/index.js
@@ -1,0 +1,9 @@
+const User = require('./User');
+const Campaign = require('./Campaign');
+const PlayerCharacter = require('./PlayerCharacter');
+
+module.exports = {
+  User,
+  Campaign,
+  PlayerCharacter,
+};
